+++ ShinyAirSensor/AirShiny/server.R
@@ -1,112 +1,175 @@
 #
-# This is the server logic of a Shiny web application. You can run the
+# This is the server logic of AirShiny web application. You can run the
 # application by clicking 'Run App' above.
 #
-# Find out more about building applications with Shiny here:
+# Find out more about Shiny applications here: 
 #
 #    http://shiny.rstudio.com/
 #
+# - Mazama Science
 
-library(AirSensor)
 # ---- Debug 
-pas <- example_pas[which(stringr::str_detect(example_pas$label, "SCNP")),] 
-pas_labels <- 
-    pas$label[-which(stringr::str_detect(pas$label, " B"))]
+library(AirSensor)
+library(rlang)
+library(magrittr)
+pas <- AirSensor::pas_load()
+pas_community <- unique(pas$communityRegion) %>% na.omit()
 # ----
 
 # Define server logic 
 shiny::shinyServer(
-    function(input, output) {
+    function(input, output, session) {
+    
+    # Leaflet render
+    output$leaflet <- 
+        leaflet::renderLeaflet({
+            
+            pas_in_comm <- 
+                pas[which(
+                    stringr::str_detect(
+                        pas$communityRegion, 
+                        input$comm_select)
+                ),]
+            
+            AirSensor::pas_leaflet_shiny(
+                pas_in_comm, 
+                parameter = "pm25_1day", 
+                paletteName = "Purple"
+            )
+            
+        })
+    
+    # Update Selected pas based on leaflet selection
+    shiny::observe({
         
-        output$pm25_plot <- 
-            shiny::renderPlot({
+        pas_in_comm <- 
+            pas[which(
+                stringr::str_detect(
+                    pas$communityRegion, 
+                    input$comm_select
+                ) & 
+                !stringr::str_detect(
+                    pas$label, 
+                    " B"
+                )
+            ),]
+        
+        shiny::updateSelectInput(
+            session, 
+            inputId = "pas_select", 
+            selected = input$leaflet_marker_click[1], 
+            choices = pas_in_comm$label
+        )
+        
+    })
+    
+    # # Text debug output 
+    #  output$test <-
+    #      shiny::renderText(paste0(input$leaflet_marker_click))
+    
+    # Standard Plot output   
+    output$selected_plot <-
+        shiny::renderPlot({
+            
+            # NOTE: The current method is not filtering ANY outliers for
+            # NOTE: ANY of the plots - may be prone to change.
+            pat <- 
+                AirSensor::pat_load(
+                    label = input$leaflet_marker_click[1], 
+                    startdate = input$date_range[1], 
+                    enddate = input$date_range[2]
+                ) %>% 
+                AirSensor::pat_outliers(replace = T, showPlot = F)
+            
+            if ( input$plot_type_select == "daily_plot" ) {
                 
-                pat <- 
-                    AirSensor::pat_load(
-                        label = input$leaflet_marker_click[1], #uses marker id from leaflet
-                        startdate = input$date_range[1],
-                        enddate = input$date_range[2]
-                    )
+                return(AirSensor::shiny_barplot(pat, period = "1 day"))
                 
-                    AirSensor::pat_multiplot(
-                        pat, 
-                        columns = 4)
-
-            })
-        
-        # Leaflet render
-        output$leaflet <- 
-            leaflet::renderLeaflet({
-                if ( stringr::str_detect(input$leaflet_select, "pm25") ) {
-                    AirSensor::pas_leaflet(
-                        pas, 
-                        parameter = input$leaflet_select, 
-                        paletteName = "Spectral")
-                }  else {
-                    AirSensor::pas_leaflet(
-                        pas,
-                        parameter = input$leaflet_select, 
-                        paletteName = input$leaflet_select
-                    )
-                }
-            })
-        
-        output$test <-
-            shiny::renderText(paste0(input$leaflet_marker_click))
-        
-        output$pm25_daily_plot <- 
-            shiny::renderPlot({
-                pat <- 
+            } else if ( input$plot_type_select == "multi_plot" ) { 
+                
+                return(AirSensor::pat_multiplot(pat))
+                
+            }  else if ( input$plot_type_select == "hourly_plot" ) {
+                
+                return(AirSensor::shiny_barplot(pat, period = "1 hour"))
+            }
+            
+        })
+    
+    # TODO: HANDLE SPECIAL DYGRAPH CASE
+    
+    # Data Table
+    output$data_explorer <- 
+        shiny::renderDataTable({
+            
+            pat <- 
+                AirSensor::pat_load(
+                    label = input$leaflet_marker_click[1], 
+                    startdate = input$date_range[1], 
+                    enddate = input$date_range[2]
+                )
+            
+            data <- pat$data 
+            
+            return(data)
+            
+        })
+    
+    # Meta Table
+    output$meta_explorer <- 
+        shiny::renderTable({
+            
+            pat <- 
+                AirSensor::pat_load(
+                    label = input$leaflet_marker_click[1], 
+                    startdate = input$date_range[1], 
+                    enddate = input$date_range[2]
+                )
+            
+            meta <-
+                dplyr::select(
+                    pat$meta,
+                    "Sensor Label" = .data$label,
+                    "Sensor Type" = .data$sensorType,
+                    "Longitude" = .data$longitude, 
+                    "Latitude" = .data$latitude,
+                    "State" = .data$stateCode, 
+                    "Country" = .data$countryCode, 
+                    "Timezone" = .data$timezone
+                )
+            
+            return(meta)
+            
+        })
+    
+    # Download button
+    output$download_data <- 
+        shiny::downloadHandler(
+            filename = function() {
+                paste0(
+                    input$leaflet_marker_click[1],
+                    "_",
+                    input$date_range[1],
+                    "_",
+                    input$date_range[2], 
+                    ".csv"
+                )
+            }, 
+            
+            content = function(file) {
+                write.csv(
                     AirSensor::pat_load(
                         label = input$leaflet_marker_click[1], 
                         startdate = input$date_range[1], 
                         enddate = input$date_range[2]
-                    ) 
-                ast <- 
-                    AirSensor::pat_createASTimeseries(
-                        pat = pat, 
-                        period = "1 day" 
-                    )
+                    )$data,
+                    file = file
+                )
                 
-                pm25_AB_avg <- 
-                    ast$data %>% 
-                    dplyr::select(
-                        .data$pm25_A_mean, 
-                        .data$pm25_B_mean
-                    ) %>% 
-                    dplyr::transmute(
-                        pm25_AB_avg = 
-                            (.data$pm25_A_mean + .data$pm25_B_mean) / 2
-                    )
-                
-                pm25_plot <- 
-                    ast$data %>% 
-                    ggplot2::ggplot(
-                        ggplot2::aes(
-                            x = .data$datetime, 
-                            y = pm25_AB_avg
-                        )
-                    ) + 
-                    ggplot2::ggtitle(
-                        label = "PM2.5"
-                    ) + 
-                    ggplot2::xlab("Datetime") + 
-                    ggplot2::ylab("\u03bcg / m\u00b3")
-                
-                pm25_avg_bar <- 
-                    ggplot2::geom_bar(                   
-                        data = ast$data,
-                        mapping = ggplot2::aes(
-                            x = .data$datetime, 
-                            y = pm25_AB_avg[,1]
-                        ), 
-                        stat = "identity"
-                    )
-                
-                return(pm25_plot + pm25_avg_bar)
-                
-            })
-        
-        
+            }
+            
+        )
+    
     }
+
 )